+++ Changelog.md
@@ -3,6 +3,7 @@
 ### 1.6 – Priority queue update
 - Added priorities in queue: added parameter "Priority" into Workflow UI. Queue runner select workflows with higher priority first. In queue page priorities are shown as separate queue. You can change maximal value of priority in the options (set 1 to hide priorities and return the old queue)
 - Added hotkey 1-9 buttons to select priority in queue page
+- Support for wildcards (`*`, `?`, `[seq]`, `[!seq]`) in hide workflows option
 
 ### 1.5 – Search update
 - Support negatives in loras and presets filter